Aula 03 - MCP II/requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
pypdf
pydantic
python-dotenv
tiktoken
mcp
langchain-mcp-adapters
httpx
httpx-sse
langgraph
